streamlit_app/requirements.txt
streamlit
pandas
matplotlib
requests
python-dotenv
numpy
